move tests to koku/koku/api/report/test/aws/tests_queries.py
